+++ src/lib/pubkey/dl_algo/dl_algo.cpp
@@ -35,7 +35,7 @@ std::vector<byte> DL_Scheme_PublicKey::public_key_bits() const
    }
 
 DL_Scheme_PublicKey::DL_Scheme_PublicKey(const AlgorithmIdentifier& alg_id,
-                                         const secure_vector<byte>& key_bits,
+                                         const std::vector<byte>& key_bits,
                                          DL_Group::Format format)
    {
    m_group.BER_decode(alg_id.parameters, format);