+++ src/lib/pubkey/gost_3410/gost_3410.h
@@ -33,10 +33,10 @@ class BOTAN_DLL GOST_3410_PublicKey : public virtual EC_PublicKey
       /**
       * Load a public key.
       * @param alg_id the X.509 algorithm identifier
-      * @param key_bits X.509 subject public key info structure
+      * @param key_bits DER encoded public key bits
       */
       GOST_3410_PublicKey(const AlgorithmIdentifier& alg_id,
-                          const secure_vector<byte>& key_bits);
+                          const std::vector<byte>& key_bits);
 
       /**
       * Get this keys algorithm name.